+++ src/MD_Parola_VScroll.cpp
@@ -58,7 +58,7 @@ void MD_PZone::effectVScroll(bool bUp, bool bIn)
       {
         uint8_t c = _MX->getColumn(j);
 
-        for (int8_t i = _nextPos; i > 0; i--)
+        for (int16_t i = _nextPos; i > 0; i--)
         {
           c = (bUp ? c << 1 : c >> 1);
           if (_inverted) c |= (bUp ? 0x01 : 0x80);
